requirements/base.txt
Django==3.0.7
django-environ==0.4.5
psycopg2==2.8.5
djangorestframework==3.11.0
django-filter==2.4.0
celery == 5.0.5
django-celery-results ==2.0.1
django-celery-beat == 2.2.0
pytest